[MoE] Assert topk_ids passed to moe_prepare_fused is torch.int32 tensor

+++ vllm_xcpu_plugin/layers/fused_moe/mpi_alltoallv_prepare_finalize_v1.py
@@ -234,7 +234,7 @@ class MpiAlltoallvPrepareAndFinalize(mk.FusedMoEPrepareAndFinalize):
             sort_indices,
             workspace,
             a1,
-            topk_ids,
+            topk_ids.to(torch.int32),
             static_buffer_size,
             num_experts,
             self.num_local_experts,
